+++ src/store/index.js
@@ -4,6 +4,10 @@ import rootReducer from './reducers/root.reducer';
 import piecesMiddleware from './middleware/pieces.middleware';
 import localStorageMiddleware from './middleware/local-storage.middleware';
 import beforeUnloadMiddleware from './middleware/before-unload.middleware';
+import silentHtml5AudioMiddleware from './middleware/silent-html5-audio.middleware';
+import mediaSessionMiddleware from './middleware/media-session.middleware';
+import shortcutsMiddleware from './middleware/shortcuts.middleware';
+import onlineStatusMiddleware from './middleware/online-status.middleware';
 import STATE_STORAGE_KEY from './middleware/local-storage.middleware/key';
 import getOnlineStatus from './get-online-status';
 import pieces from '../pieces/index';
@@ -32,7 +36,14 @@ if (isMobile) {
   initialState.isMuted = false;
 }
 
-const allMiddlewares = [piecesMiddleware, localStorageMiddleware];
+const allMiddlewares = [
+  piecesMiddleware,
+  silentHtml5AudioMiddleware,
+  mediaSessionMiddleware,
+  shortcutsMiddleware,
+  onlineStatusMiddleware,
+  localStorageMiddleware,
+];
 const desktopMiddlewares = allMiddlewares.concat([beforeUnloadMiddleware]);
 
 const store = createStore(